refactor(app): move QuickTransfer pages
