+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -34,8 +34,8 @@ import { RobotSettingsDashboard } from '../pages/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../pages/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
 import { QuickTransferFlow } from '../organisms/QuickTransferFlow'
-import { QuickTransferDashboard } from '../pages/QuickTransferDashboard'
-import { QuickTransferDetails } from '../pages/QuickTransferDetails'
+import { QuickTransferDashboard } from '../pages/ODD/QuickTransferDashboard'
+import { QuickTransferDetails } from '../pages/ODD/QuickTransferDetails'
 import { RunningProtocol } from '../pages/RunningProtocol'
 import { RunSummary } from '../pages/RunSummary'
 import { UpdateRobot } from '../pages/UpdateRobot/UpdateRobot'
