+++ app/src/pages/ODD/QuickTransferDashboard/IntroductoryModal.tsx
@@ -8,10 +8,10 @@ import {
   ALIGN_CENTER,
   TEXT_ALIGN_CENTER,
 } from '@opentrons/components'
-import { OddModal } from '../../molecules/OddModal'
-import { SmallButton } from '../../atoms/buttons'
+import { OddModal } from '../../../molecules/OddModal'
+import { SmallButton } from '../../../atoms/buttons'
 
-import imgSrc from '../../assets/images/on-device-display/odd-abstract-6.png'
+import imgSrc from '../../../assets/images/on-device-display/odd-abstract-6.png'
 
 interface IntroductoryModalProps {
   onClose: () => void
