+++ app/src/pages/ODD/QuickTransferDashboard/NoQuickTransfers.tsx
@@ -12,7 +12,7 @@ import {
   StyledText,
 } from '@opentrons/components'
 
-import imgSrc from '../../assets/images/on-device-display/empty_quick_transfer_dashboard.png'
+import imgSrc from '../../../assets/images/on-device-display/empty_quick_transfer_dashboard.png'
 
 export function NoQuickTransfers(): JSX.Element {
   const { t } = useTranslation('quick_transfer')
