+++ app/src/pages/ODD/QuickTransferDashboard/__tests__/LongPressModal.test.tsx
@@ -7,8 +7,8 @@ import { fireEvent, renderHook, screen } from '@testing-library/react'
 import { useLongPress } from '@opentrons/components'
 import { useCreateRunMutation, useHost } from '@opentrons/react-api-client'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 import { LongPressModal } from '../LongPressModal'
 
 import type { HostConfig } from '@opentrons/api-client'
