+++ app/src/pages/ODD/QuickTransferDashboard/__tests__/QuickTransferCard.test.tsx
@@ -7,8 +7,8 @@ import {
   useProtocolAnalysisAsDocumentQuery,
 } from '@opentrons/react-api-client'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 import { QuickTransferCard } from '../QuickTransferCard'
 import { LongPressModal } from '../LongPressModal'
 import type { NavigateFunction } from 'react-router-dom'
