+++ app/src/pages/ODD/QuickTransferDashboard/index.tsx
@@ -21,14 +21,14 @@ import {
   useInstrumentsQuery,
 } from '@opentrons/react-api-client'
 
-import { SmallButton, FloatingActionButton } from '../../atoms/buttons'
-import { Navigation } from '../../organisms/Navigation'
+import { SmallButton, FloatingActionButton } from '../../../atoms/buttons'
+import { Navigation } from '../../../organisms/Navigation'
 import {
   getPinnedQuickTransferIds,
   getQuickTransfersOnDeviceSortKey,
   getHasDismissedQuickTransferIntro,
   updateConfigValue,
-} from '../../redux/config'
+} from '../../../redux/config'
 import { PinnedTransferCarousel } from './PinnedTransferCarousel'
 import { sortQuickTransfers } from './utils'
 import { QuickTransferCard } from './QuickTransferCard'
@@ -40,8 +40,8 @@ import { DeleteTransferConfirmationModal } from './DeleteTransferConfirmationMod
 
 import type { ProtocolResource } from '@opentrons/shared-data'
 import type { PipetteData } from '@opentrons/api-client'
-import type { Dispatch } from '../../redux/types'
-import type { QuickTransfersOnDeviceSortKey } from '../../redux/config/types'
+import type { Dispatch } from '../../../redux/types'
+import type { QuickTransfersOnDeviceSortKey } from '../../../redux/config/types'
 
 export function QuickTransferDashboard(): JSX.Element {
   const protocols = useAllProtocolsQuery()
