+++ app/src/pages/ODD/QuickTransferDetails/Deck.tsx
