+++ app/src/pages/ODD/QuickTransferDetails/Hardware.tsx
@@ -26,13 +26,13 @@ import {
 import {
   useGripperDisplayName,
   usePipetteNameSpecs,
-} from '../../resources/instruments/hooks'
-import { useRequiredProtocolHardware } from '../../pages/Desktop/Protocols/hooks'
+} from '../../../resources/instruments/hooks'
+import { useRequiredProtocolHardware } from '../../../pages/Desktop/Protocols/hooks'
 
 import type {
   ProtocolHardware,
   ProtocolPipette,
-} from '../../pages/Desktop/Protocols/hooks'
+} from '../../../pages/Desktop/Protocols/hooks'
 import type { TFunction } from 'i18next'
 
 const Table = styled('table')`
