+++ app/src/pages/ODD/QuickTransferDetails/__tests__/Deck.test.tsx
@@ -3,13 +3,13 @@ import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 import { screen } from '@testing-library/react'
 import { when } from 'vitest-when'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 import {
   useProtocolAnalysisAsDocumentQuery,
   useProtocolQuery,
 } from '@opentrons/react-api-client'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { Deck } from '../Deck'
 
 import type { UseQueryResult } from 'react-query'
